fix(ios): toLong() cast for createProxyHttpClient timeout params, bump build to 392

+++ YPtun/sharedUI/src/iosMain/kotlin/org/olcbox/app/vpn/IosVpnManager.kt
@@ -368,9 +368,9 @@ class IosVpnManager(
 
     private val pingClient: HttpClient by lazy {
         createProxyHttpClient(
-            connectTimeoutMs = PING_TIMEOUT_MS,
-            requestTimeoutMs = PING_TIMEOUT_MS,
-            socketTimeoutMs = PING_TIMEOUT_MS,
+            connectTimeoutMs = PING_TIMEOUT_MS.toLong(),
+            requestTimeoutMs = PING_TIMEOUT_MS.toLong(),
+            socketTimeoutMs = PING_TIMEOUT_MS.toLong(),
         )
     }
